oracle-backend/requirements.txt
web3==6.11.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
